feat: add team tooltip to rankings
